+++ packages/xrpl/src/models/ledger/index.ts
@@ -2,16 +2,20 @@ import AccountRoot, {
   AccountRootFlags,
   AccountRootFlagsInterface,
 } from './AccountRoot'
-import Amendments from './Amendments'
+import Amendments, { Majority, AMENDMENTS_ID } from './Amendments'
 import Check from './Check'
 import DepositPreauth from './DepositPreauth'
 import DirectoryNode from './DirectoryNode'
 import Escrow from './Escrow'
-import FeeSettings from './FeeSettings'
+import FeeSettings, {
+  FeeSettingsPreAmendmentFields,
+  FeeSettingsPostAmendmentFields,
+  FEE_SETTINGS_ID,
+} from './FeeSettings'
 import Ledger from './Ledger'
 import LedgerEntry from './LedgerEntry'
 import LedgerHashes from './LedgerHashes'
-import NegativeUNL from './NegativeUNL'
+import NegativeUNL, { NEGATIVE_UNL_ID } from './NegativeUNL'
 import { NFTokenOffer } from './NFTokenOffer'
 import { NFToken, NFTokenPage } from './NFTokenPage'
 import Offer, { OfferFlags } from './Offer'
@@ -24,15 +28,21 @@ export {
   AccountRoot,
   AccountRootFlags,
   AccountRootFlagsInterface,
+  AMENDMENTS_ID,
   Amendments,
   Check,
   DepositPreauth,
   DirectoryNode,
   Escrow,
+  FEE_SETTINGS_ID,
   FeeSettings,
+  FeeSettingsPreAmendmentFields,
+  FeeSettingsPostAmendmentFields,
   Ledger,
   LedgerEntry,
   LedgerHashes,
+  Majority,
+  NEGATIVE_UNL_ID,
   NegativeUNL,
   NFTokenOffer,
   NFTokenPage,