+++ packages/xrpl/src/models/transactions/index.ts
@@ -1,5 +1,6 @@
+export { BaseTransaction } from './common'
 export { validate, TransactionAndMetadata, Transaction } from './transaction'
-export { TransactionMetadata } from './metadata'
+export * from './metadata'
 export {
   AccountSetAsfFlags,
   AccountSetTfFlags,
@@ -18,6 +19,7 @@ export {
 export { EscrowCancel } from './escrowCancel'
 export { EscrowCreate } from './escrowCreate'
 export { EscrowFinish } from './escrowFinish'
+export { EnableAmendment, EnableAmendmentFlags } from './enableAmendment'
 export { NFTokenAcceptOffer } from './NFTokenAcceptOffer'
 export { NFTokenBurn } from './NFTokenBurn'
 export { NFTokenCancelOffer } from './NFTokenCancelOffer'
@@ -45,7 +47,9 @@ export {
 } from './paymentChannelClaim'
 export { PaymentChannelCreate } from './paymentChannelCreate'
 export { PaymentChannelFund } from './paymentChannelFund'
+export { SetFee, SetFeePreAmendment, SetFeePostAmendment } from './setFee'
 export { SetRegularKey } from './setRegularKey'
 export { SignerListSet } from './signerListSet'
 export { TicketCreate } from './ticketCreate'
 export { TrustSetFlagsInterface, TrustSetFlags, TrustSet } from './trustSet'
+export { UNLModify } from './UNLModify'