+++ packages/xrpl/HISTORY.md
@@ -3,7 +3,7 @@
 Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xrpl-announce) for release announcements. We recommend that xrpl.js (ripple-lib) users stay up-to-date with the latest stable release.
 ## Unreleased
 
-## Added
+### Added
 * Add `BurnedNFTokens`, `FirstNFTSequence`, `MintedNFTokens`,
 `NFTokenMinter`, and `WalletLocator` to `AccountRoot`.
 * Add `ledger_hash` and `ledger_index` to `account_nfts`,
@@ -12,6 +12,17 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 * Add types for `NFTokenPage` and `NFTokenOffer` LedgerEntries.
 * Add type for NFToken object that is stored on a `NFTokenPage`.
 * Add type for `account_info`'s `account_flags` property.
+* Add types for `EnableAmendment`, `SetFee`, and `UNLModify` transactions.
+* Add the new fields for `XRPFees` amendment and id for the `FeeSettings`
+* Add `FeeSettings`, `NegativeUNL`, and `Amendments` singleton ledger entry ids.
+* Add `WalletLocator` to `SignerEntry` on `SignerList` (LedgerEntry).
+* Export many nested types and interfaces
+
+### Breaking
+* If you were deep importing these types previously you will need to import them from `xrpl` and rename them:
+  * `methods/accountLines`: `Trustline` -> `AccountLinesTrustline`
+  * `methods/bookOffers`: `TakerAmount` -> `BookOfferCurrency`
+  * `methods/ledgerData`: `BinaryLedgerEntry` -> `LedgerDataBinaryLedgerEntry`
 
 ### Changed
 * `DeposityPreauth` has a new type definition that will not allow `Authorize` and `Unauthorize` on the same instance. This provides better guard support.